+++ cmake_lint/lintdb.py
@@ -256,5 +256,8 @@ and you probably aren't assigning to the correct variable.
 This warning may be emitted for assignment (e.g. `set()` or `list()`) as
 well as for variable expansion in an argument (e.g. `"${CMAKE_Cxx_STANDARD}"`).
 """
+}), (
+"W0106",
+"String looks like a variable reference missing an {:s} tag '{:s}'", {
 })
 ]